Unhardcoded the file location, removed a useless function and general restructure

+++ src/main/java/com/accenture/repository/BookRepository.java
@@ -1,9 +1,9 @@
 package com.accenture.repository;
 
-import com.accenture.modules.book.Book;
+import com.accenture.modules.Book;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookRepository extends CrudRepository<Book, Long>{
+public interface BookRepository extends CrudRepository<Book, Integer>{
 }
