+++ src/main/java/com/accenture/repository/CDRepository.java
@@ -1,9 +1,9 @@
 package com.accenture.repository;
 
-import com.accenture.modules.cd.CD;
+import com.accenture.modules.CD;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CDRepository extends CrudRepository<CD, Long> {
+public interface CDRepository extends CrudRepository<CD, Integer> {
 }
