+++ src/main/java/com/accenture/repository/ItemRepository.java
@@ -1,9 +1,9 @@
 package com.accenture.repository;
 
-import com.accenture.modules.item.Item;
+import com.accenture.modules.Item;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemRepository extends CrudRepository<Item, Long> {
+public interface ItemRepository extends CrudRepository<Item, Integer> {
 }
